tests(auth): consolidate authRoute tests
