+++ src/v2/Apps/Authentication/authenticationRoutes.tsx
@@ -68,8 +68,6 @@ export const authenticationRoutes: AppRouteConfig[] = [
     },
     onClientSideRender: ({ match }) => {
       setCookies(match.location.query)
-
-      // TODO: check if user is logged in and redirect
       LoginRoute.preload()
     },
   },
@@ -85,16 +83,10 @@ export const authenticationRoutes: AppRouteConfig[] = [
     hideFooter: true,
     getComponent: () => ResetPasswordRoute,
     onServerSideRender: ({ req, res }) => {
-      if (req.query.reset_password_token) {
-        req.session.reset_password_token = req.query.reset_password_token
-        req.session.set_password = req.query.set_password
-        req.session.reset_password_redirect_to =
-          req.query.reset_password_redirect_to
-
-        runAuthMiddleware({ req, res })
-      } else {
+      if (!req.query.reset_password_token) {
         res.redirect("/")
       }
+      runAuthMiddleware({ req, res })
     },
     onClientSideRender: ({ match }) => {
       setCookies(match.location.query)
